feat: update bagitprofile parsing for 1.4.0 version
