+++ src/main/resources/MessageBundle.properties
@@ -2,6 +2,8 @@
 
 #for BagitProfileDeserializer.java
 fetch_allowed=Are fetch files allowed? [{}]
+fetch_required=Are fetch files required? [{}]
+data_empty=Is the data directory required to be empty? [{}]
 serialization_allowed=Serialization is: [{}]
 parsing_bagit_profile_info_section=Parsing the BagIt-Profile-Info section
 parsing_mandatory_tags_of_bagit_profile_info_section=Parsing mandatory tags of the BagIt-Profile-Info section
