+++ src/test/resources/bagitProfiles/OnlyRequiredFieldsProfile_v1.4.0.json
@@ -1,0 +1,12 @@
+{
+  "BagIt-Profile-Info": {
+    "BagIt-Profile-Identifier": "http://canadiana.org/standards/bagit/tdr_ingest.json",
+    "Source-Organization": "Candiana.org",
+    "External-Description": "BagIt profile for ingesting content into the C.O. TDR loading dock.",
+    "Version": "1.2",
+    "BagIt-Profile-Version": "1.4.0"
+  },
+  "Accept-BagIt-Version": [
+    "0.96"
+  ]
+}
